Reorganize GHC heap-related modules
