+++ test/test.hs
@@ -6,7 +6,7 @@ where
 import Cardano.Prelude
 import Test.Cardano.Prelude
 
-import qualified Test.Cardano.Prelude.GHC.Heap
+import qualified Test.Cardano.Prelude.GHC.Heap.NormalForm
 
 main :: IO ()
-main = runTests [Test.Cardano.Prelude.GHC.Heap.tests]
+main = runTests [Test.Cardano.Prelude.GHC.Heap.NormalForm.tests]
